Enable bark-label-view split/add on OS X (#71)
Fix #44 .

`Ctrl+click` is difficult to get matplotlib backends to recognize on Macs. Some don't register any key presses during mouse clicks, and others don't register `Ctrl+click`. Recognition of `Cmd+click` is inconsistent as well.

For this reason, **for Mac only**, the control sequences for splitting an interval and adding a new interval in `bark-label-view` have been changed to `Shift+click`. They remain `Ctrl+click` for linux and windows.

In the process of investigating this, I discovered that the built-in `TkAgg` backend sometimes doesn't recognize `Shift+click` the first time it is produced; it does recognize subsequent attempts. This can be a pain point in some workflows, so I added a check for the PyQt5 backend `Qt5Agg`, which doesn't have this problem, falling back on `TkAgg` if PyQt5 isn't installed. It's a large library to have as a dependency for just this feature, so I have labeled it "optional" in the installation section. As I note there, `bark-label-view` is still usable under the `TkAgg` backend, just occasionally slightly less smooth.